Conversión de toInt a toDouble

+++ app/src/main/java/com/pmdm/actividad02pmdm/Calc.kt
@@ -3,8 +3,8 @@ package com.pmdm.actividad02pmdm
 /**
  * En la clase @calc vamos a realizar las diferentes operaciones disponibles en la calculadora.
  */
-class Calc (val n1:Int, val n2:Int, val op:String){
-    fun operation(op:String):Int{
+class Calc (val n1:Double, val n2:Double, val op:String){
+    fun operation(op:String):Double{
         /**
          * Dentro del método @operation de la clase @calc pondremos los
          * diferentes cálculos que la calculadora podrá realizar. Usando el @return when
@@ -12,11 +12,11 @@ class Calc (val n1:Int, val n2:Int, val op:String){
          * que contenga la variable @op.
          */
         return when (op) {
-            "+" -> {(n1 + n2).toInt()} // Operación suma
-            "-" -> {(n1 - n2).toInt()} // Operación resta
-            "*" -> {(n1 * n2).toInt()} // Operación Multiplicación
-            "/" -> {(n1 / n2).toInt()} // Operación División
-            else -> 0
+            "+" -> {(n1 + n2).toDouble()} // Operación suma
+            "-" -> {(n1 - n2).toDouble()} // Operación resta
+            "*" -> {(n1 * n2).toDouble()} // Operación Multiplicación
+            "/" -> {(n1 / n2).toDouble()} // Operación División
+            else -> 0.0
         }
     }
 }
